+++ src/jk_logging/fmt/JSONLogMessageFormatter.py
@@ -60,13 +60,13 @@ class JSONLogMessageFormatter(AbstractLogMessageFormatter):
 			self.__indentBuffer += self.__fillChar
 		sIndent = self.__indentBuffer[0:indentationLevel]
 		sParentID = str(logEntryStruct[3]) if (logEntryStruct != None) else "-"
-		sTimeStamp = "[" + self.__timeStampFormatter(logEntryStruct[4]) + "]"
+		sTimeStamp = self.__timeStampFormatter(logEntryStruct[4])
 		sLogType = AbstractLogMessageFormatter.LOG_LEVEL_TO_STR_MAP__RIGHT_ALIGNED[logEntryStruct[5]]
 
 		s = sIndent
 		if self.__includeIDs:
 			s += "(" + sParentID + "|" + sID + ") "
-		s += sTimeStamp + " "
+		s += "[" + sTimeStamp + "] "
 
 		if logEntryStruct[0] == "txt":
 			sLogMsg = logEntryStruct[6]